+++ tests/test_cli.py
@@ -11,16 +11,21 @@ import pandas as pd
 
 
 def test_jvm_warn(recwarn):
-    """Test for a JVM start-up warning.
+    """Test that no warnings are issued on JVM start-up.
 
-    This will be emitted, e.g. for JPype 0.7 if the 'convertStrings' kwarg is
-    not provided to jpype.startJVM.
+    A warning message is emitted e.g. for JPype 0.7 if the 'convertStrings'
+    kwarg is not provided to jpype.startJVM.
 
-    NB this should be in test_core.py, but because pytest executes tests in
-    file, then code order, it must be before the call to ix.Platform() below.
+    NB this function should be in test_core.py, but because pytest executes
+    tests in file, then code order, it must be before the call to ix.Platform()
+    below.
     """
+
+    # Start the JVM for the first time in the test session
     ix.start_jvm()
+
     if jpype.__version__ > '0.7':
+        # Zero warnings were recorded
         assert len(recwarn) == 0, recwarn.pop().message